# openapi-spec.yaml
openapi: 3.0.0
info:
  title: NightGuide API
  version: 0.1.0
servers:
  - url: https://api.nightguide.app
  - url: https://api.dev.nightguide.app
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      parameters:
        - in: query
          name: ids
          schema:
            type: array
            items:
              type: string
        - in: query
          name: status
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Result orders.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  totalCount:
                    type: integer
                    format: int32
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create a order.
      operationId: createOrder
      tags:
        - orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/{orderId}:
    get:
      summary: Get a order.
      operationId: getOrder
      tags:
        - orders
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: Order ID.
      responses:
        '200':
          description: Order data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      summary: Update a order.
      operationId: updateOrder
      tags:
        - orders
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: Order ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a order.
      operationId: deleteOrder
      tags:
        - orders
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: Order ID.
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/{orderId}/metadata/{metaKey}:
    put:
      summary: Update order metadata.
      operationId: updateOrderMetadata
      tags:
        - orders
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: Order ID.
        - in: path
          name: metaKey
          schema:
            type: string
          required: true
          description: Metadata key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  $ref: '#/components/schemas/AnyValue'
      responses:
        '200':
          description: Order metadata.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /orders/{orderId}/downloads:
    get:
      summary: Get order downloads.
      operationId: getOrderDownloads
      tags:
        - orders
      parameters:
        - in: path
          name: orderId
          schema:
            type: string
          required: true
          description: Order ID.
        - in: query
          name: key
          schema:
            type: string
          required: true
          description: Download key
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  responses:
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: The request did not satisfy the requirements
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Success:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
            required:
              - success
  schemas:
...
    Order:
      type: object
      properties:
        status:
          type: string
        billingDetails:
          type: object
          properties:
            email:
              type: string
            name:
              type: string
            address:
              type: object
              properties:
                city:
                  type: string
                country:
                  type: string
                line1:
                  type: string
                line2:
                  type: string
                postalCode:
                  type: string
                state:
                  type: string
            items:
              type: array
              items:
                type: object
                properties:
                  sku:
                    type: string
                  quantity:
                    type: integer
                    format: int32
                    minimum: 1
                  price:
                    type: number
                    format: float
                    minimum: 0
                  currency:
                    type: string
            metadata:
              type: object
...
    Error:
      type: object
      properties:
        status:
          type: number
          description: HTTP status code
        type:
          type: string
          description: Error identifier
        message:
          type: string
          description: Human-readable error description
      required:
        - status
        - type
    AnyValue: {}